add missing newlines to content blocks of gbs (#1394)
## Summary

This PR adds some missing line breaks between reStructuredText sections
of the [`gbs` demo](https://pennylane.ai/qml/demos/gbs). Also updates
the hash lines to have consistent length of 79, as specified in the
README.

Currently, the lack of newline above some of the sections in this demo
causes sphinx to incorrectly build the final HTML output. This results
in partially or fully missing paragraphs in the final rendered page.

This only affects a few of the sections of the `gbs` demo.