Update workflows and deploy scripts to use `pr_number` as preview argument (#1400)
## Summary

This pull request updates the deployment workflows to replace the
"preview" flag with a "pr_number" parameter, streamlining the process of
deploying demos for pull request previews and main branch deployments.
The changes enhance clarity and flexibility by explicitly using pull
request numbers to distinguish between deployment types.

### Deployment Workflow Updates:

*
[`.github/workflows/qml_pipeline_v2/deploy.py`](diffhunk://#diff-8d18985df6c6365fc1e13492f70546e1caf35ee779a91f6423a41221e7c52c29L13-R13):
Replaced the `--preview` argument with `--pr-number`, updated the
deployment URL template to use `pr_number`, and adjusted the logic to
handle pull request numbers for deployment.
[[1]](diffhunk://#diff-8d18985df6c6365fc1e13492f70546e1caf35ee779a91f6423a41221e7c52c29L13-R13)
[[2]](diffhunk://#diff-8d18985df6c6365fc1e13492f70546e1caf35ee779a91f6423a41221e7c52c29L24-R34)
[[3]](diffhunk://#diff-8d18985df6c6365fc1e13492f70546e1caf35ee779a91f6423a41221e7c52c29L53-R52)

*
[`.github/workflows/v2-build-and-deploy.yml`](diffhunk://#diff-924e8738afa8b3a1299cd57b68a91d294dd95d10fb2d7968c253012a83e32b7cL31-R34):
Removed the `as-previews` input and added a `pr_number` input to specify
the pull request number for deployments. Updated job references to use
`pr_number` instead of `as-previews`.
[[1]](diffhunk://#diff-924e8738afa8b3a1299cd57b68a91d294dd95d10fb2d7968c253012a83e32b7cL31-R34)
[[2]](diffhunk://#diff-924e8738afa8b3a1299cd57b68a91d294dd95d10fb2d7968c253012a83e32b7cL102-R97)

### Additional Workflow Adjustments:

*
[`.github/workflows/v2-deploy-demos.yml`](diffhunk://#diff-3ba5ae40c65bc1849456f58c31ccf2b13b308a3ee5b6c349348c44fd39381365L16-R22):
Commented out the `preview` input and replaced it with `pr_number` for
deployment configuration. Updated the deployment command to use
`--pr-number`.
[[1]](diffhunk://#diff-3ba5ae40c65bc1849456f58c31ccf2b13b308a3ee5b6c349348c44fd39381365L16-R22)
[[2]](diffhunk://#diff-3ba5ae40c65bc1849456f58c31ccf2b13b308a3ee5b6c349348c44fd39381365L34-R44)
[[3]](diffhunk://#diff-3ba5ae40c65bc1849456f58c31ccf2b13b308a3ee5b6c349348c44fd39381365L77-R85)

*
[`.github/workflows/v2-deploy-pr.yml`](diffhunk://#diff-bb1f887f50cfc7839e56ce823ccf5b9b7a4a07bfd475c5646b28105610511bf3L87-R90):
Changed the deployment environment from `swc-staging` to `swc-dev` and
replaced the `preview` flag with `pr_number` in job inputs and comments.
[[1]](diffhunk://#diff-bb1f887f50cfc7839e56ce823ccf5b9b7a4a07bfd475c5646b28105610511bf3L87-R90)
[[2]](diffhunk://#diff-bb1f887f50cfc7839e56ce823ccf5b9b7a4a07bfd475c5646b28105610511bf3L112-R112)