Update deprecated parameters in pyproject and bump numpy (#1414)
**Title:** Update deprecated parameters in pyproject and pin numpyro

**Summary:**
The current master builds of QML are failing, from inspection of the
logs the main culprit seems to be the version of numpyro being bumped.

Upon further inspection of the lockfile, it seems that the lockfile was
also re-generated using an older version of poetry than what QML is
using in CI. We use 2.x within CI but the lockfile was generated with
1.x.

Though poetry 2.x tries it's best to be compatible with 1.x lockfile and
pyproject, it is not in our best interest to support both, as it can
cause problems.

As such, this PR updates the pyproject to address all deprecation
warnings from poetry 2.x, including adding a minimum poetry version.
This pyproject is now incompatible with poetry 1.x, and using poetry 1.x
locally with QML repo will yield an error now.

**Relevant references:**
sc-93662

**Possible Drawbacks:**
Users using poetry 1.x locally will need to use 2.x when updating
lockfile for QML.

**Related GitHub Issues:**
None.