QML V2 dev branch fixes (#1423)
This PR introduces a few fixes to the V2 demos pipeline to get all demos
to build:
- Add support for pennylane v0.41.1 and v0.42.0 wrt the
`PennyLaneDeprecationWarning` since the containing module has been
refactored. This warning is imported by Sphinx before the V2 pipeline
installs the dev version of pennylane, so we need a try/catch for
importing this module.
- Several changes to the dev constraints file, including cirq and
constraining PennyLane to install from the RC branch.
- A big change to the `qml build` tool. The current tool assumes all
constraints can be satisfied together, which just isn't possible once we
start installing the `dev` versions of PennyLane, Catalyst, and several
plugins. This is addressed here by adding an additional step to the `qml
build` tool to install each of these libraries separately at their
correct versions. This step mirrors what is currently done in the
Makefile.

The above have been tested and verified that the entire V2 dev pipeline
builds successfully

![Screenshot 2025-07-09 at 11 15
29 AM](https://github.com/user-attachments/assets/8440aca4-b852-4898-b585-ba5f79315159)

This PR also syncs recent changes to the demos in `dev` into the
`demonstrations_v2` folder, and increases the verbosity of the `qml`
build tool so we have better visibility into the behaviour of the
pipeline while we're still in testing mode.