Update `dev` and `master` v2 workflows (#1432)
The `master` and `dev` branch builds are used for two purposes:
- As a canary to surface potential issues introduced into the demos via
changes to the underlying dependencies
- In the [qml-diff-checker
workflow](https://github.com/PennyLaneAI/qml-diff-checker) to compare
the two builds priori to a new release of PennyLane.

This PR adjusts the build frequency of these pipelines, as well as the
retention period of the resulting build artifacts:
- Two build were occurring over the weekend, which is redundant as the
odds of a material change between Saturday and Sunday is low, and no one
is likely to action it anyway during that time. The new cron frequency
builds these pipelines Monday morning, Wednesday morning, and Friday
morning at 12am so any changes are apparent at the start of the work
day.
- The retention period of the consolidated artifacts is bumped from 1
day (default) to 3 days for these pipelines, so that there is always a
fresh artifact associated with a successful run of these pipelines. This
is required for the [qml-diff-checker
workflow](https://github.com/PennyLaneAI/qml-diff-checker).