Mark `tutorial_error_mitigation` as non-executable (#1465)
The `tutorial_error_mitigation.py` has to be marked as non-executable to
unblock the `dev` branch CI.

The issue is related to [this
PR](https://github.com/PennyLaneAI/pennylane/pull/7909) in PennyLane
which is introducing the deprecation of the `QuantumScript.to_openqasm`
method in v0.43 (removal in v0.44). This method is still used in Mitiq's
`from_pennylane` conversion function (see
[here](https://github.com/unitaryfoundation/mitiq/blob/main/mitiq/interface/mitiq_pennylane/conversions.py#L28))
and it works fine for them because Mitiq currently has an explicit
dependency on PennyLane 0.36.

We will be able to mark the tutorial executable again when they update
their code to work with `pennylane>=0.43`.

---------

Co-authored-by: Alex Preciado <alex.preciado@xanadu.ai>