Optionally override demo execution via the metadata (#1467)
This PR adds the (optional) `executable` field to the demo metadata
schema, and modifies the `qml` tool to use this metadata field for
executing a demo, falling back to the `tutorial_` naming convention if
the field doesn't exist. It also modifies the Sphinx configuration to
process any `.py` file rather than only those that start with
`tutorial_`, so we can (optionally) execute demos that don't start with
`tutorial_`.