Update deploy env and commenting (#1480)
This pull request updates deployment workflows to align with the new SWC
environment structure and re-enables automated PR comments for demo
previews. The changes primarily focus on transitioning deployments to
production environments and enhancing the developer experience with
improved PR feedback.

### Deployment environment updates:
*
[`.github/workflows/v2-deploy-demos-swc-env.yml`](diffhunk://#diff-f718153c08f2a034e3c076c84a3978e1c4bfd14d7033543749ca0c78cd6e6f50L72-R75):
Updated the SWC environment mapping logic to deploy the `master` branch
to `swc-prod` instead of `swc-staging`, and removed references to
deprecated branches (`test-v2-master` and `test-v2-dev`).
*
[`.github/workflows/v2-deploy-pr.yml`](diffhunk://#diff-bb1f887f50cfc7839e56ce823ccf5b9b7a4a07bfd475c5646b28105610511bf3L88-R127):
Changed the default SWC environment for PR deployments from
`swc-staging` to `swc-prod`.

### Re-enabling PR comments:
*
[`.github/workflows/v2-deploy-pr.yml`](diffhunk://#diff-bb1f887f50cfc7839e56ce823ccf5b9b7a4a07bfd475c5646b28105610511bf3L88-R127):
Re-enabled the `generate-comment` step to create and post markdown
comments on PRs with demo preview links. The comments now include a link
to the demos landing page and a list of updated demos.