Executable changes (#1484)
This PR introduces two booleans to the metadata to control execution of
the demos:
- `executable_stable`: Whether or not this demo should be considered
executable using the most recent (stable) version of PennyLane. This
corresponds with the master branch of qml.
- `executable_latest`: Whether or not this demo should be considered
executable using the latest dev version of PennyLane. This corresponds
with the dev branch of qml.

As with the previous version, if either of these variables are missing
then the qml tool will fall back to using the presence of the
`tutorial_` prefix to determine if the demo is executable.

---------

Co-authored-by: Alan Martin <53958929+Alan-eMartin@users.noreply.github.com>