Fix notebook converter (#1360)
Update notebook converter to remove setting author info. Fix cell_id
……issue.

**Title:**
Fix QML notebook converter

**Summary:**
The current QML converter is not working with newer jupyter notebooks
due to them not having an id for each cell.

In addition, the way demo author information is stored in QML today is
completely different to how it was stored back when this converter was
first written.

This PR updates the notebook converter such that the author info is no
longer needed, and the cell_id is used only if present, if missing it
simply uses the cell number instead.

**Relevant references:**
[sc-90046]

**Possible Drawbacks:**
As per before, the generated demo can be passed to sphinx-build, which
has been tested. But it likely will not be 100% up to standard of the
Product team, but this converter should give the author a quick
conversion base to work off of.

**Related GitHub Issues:**
None.

---------

Co-authored-by: Catalina Albornoz <albornoz.catalina@hotmail.com>
Co-authored-by: Ashish Kanwar Singh <104938869+ashishks0522@users.noreply.github.com>