Custom demo rst processing (#1546)
This PR introduces a series of [Pandoc
filters](https://pandoc.org/filters.html) to customize processing of
various rst directives for the markdown output in the rendered Jupyter
Notebook versions of the demos:
- `filter_figures`
This filter strips the local directory for a figure and adds the
appropriate cloud URL where the figure will be sourced once the demo is
uploaded. For testing we've added a figure to the blog server so this
will render properly. This should also work for any demo that has
already been deployed to the website.
- `filter_directives`
This filter places previously unhandled rst directives (notes, warnings,
tips, etc.) inside a BlockQuote block. We can potentially make this
fancier in the future, but for now this is functional and looks nicer
than the previously unprocessed blocks. We also remove the \:meta\: and
\:related\: blocks from the top of the demo.

**Update:** We also now process references in this filter. Pandoc can't
handle references that occur in separate code blocks, so instead we
replace all references with a link to the references section in the demo
on pennylane.ai, and remove the (empty) "References" section at the end
of the demo.
- `filter_links`
This filter replaces the intersphinx mapping functionality, which
doesn't seem to work for the Jupyter Notebook text cells. Since it's an
attempt to reproduce the Intersphinx mapping manually, it's not perfect,
but covers the vast majority of Intersphinx reference cases. Because we
need the `objects.inv` files available locally for the mapping, this
filter does slow down the build, but it shouldn't be too noticeable.

This PR also includes a new "kitchen sink" demo, where we've added as
many rst directives as we can find in the existing demos. Processing
this demo with the qml tool should produce the expected output in the
resulting Jupyter Notebook. If new cases that are not covered here are
found, please let the @PennyLaneAI/cloud-software team know and we'll
look to get them handled accordingly.