Fix for objects.inv (#1552)
This PR tweaks the `qml` build tool such that the generated objects.inv
files match the naming and uris for the demos on pennylane.ai. This PR
should fix #1550.

Note: This has been tested locally. ~We should verify it works when
building all demos as wel.~ Tested for all demos successfully
[here](https://github.com/PennyLaneAI/qml/actions/runs/17866805882)

---------

Co-authored-by: Alan Martin <alan.martin@xanadu.ai>