merge `dev` into `master` for v0.43.0 release (#1573)
Following the release guide,
> Create a new branch from dev, and open a PR to merge the new branch
into master. Once merged in, ensure that master subsequently builds
successfully.

This merges all changes from `dev` that occured for the release of
`v0.43.0` PL ecosystem.

PR is a follow up to https://github.com/PennyLaneAI/qml/pull/1572 where
we updated dependencies in `dev`. This PR was created with,
- Make branch after pulling latest `dev`
- Create PR targeting master

⚠️ Manually updated the `constraints-stable.txt` to match the
`constraints-dev.txt` so that they build correctly. Also had to manually
update all of the metadata so that the date of last modification is
today.

⚠️ This branch was tested against all demos here:
https://github.com/PennyLaneAI/qml/actions/runs/18546834983

---------

Co-authored-by: Christina Lee <chrissie.c.l@gmail.com>
Co-authored-by: Yushao Chen (Jerry) <chenys13@outlook.com>
Co-authored-by: GitHub Nightly Merge Action <actions@github.com>
Co-authored-by: Simone Gasperini <simone.gasperini4@unibo.it>
Co-authored-by: Alex Preciado <alex.preciado@xanadu.ai>
Co-authored-by: Alan Martin <alan.martin@xanadu.ai>
Co-authored-by: Alan Martin <53958929+Alan-eMartin@users.noreply.github.com>
Co-authored-by: Isaac De Vlugt <isaacdevlugt@gmail.com>
Co-authored-by: Isaac De Vlugt <34751083+isaacdevlugt@users.noreply.github.com>
Co-authored-by: Paul Finlay <50180049+doctorperceptron@users.noreply.github.com>