Handle new sphx-gallery format (#1599)
After the update to Sphinx, Sphinx Gallery now outputs a different
format for generated code blocks, where the code block is now part of an
entirely separate element and no longer nested in `rst-class` directive.
This PR adds a fix to be able to handle both the new and old
`sphx-glr-script-out` formats. It has been tested on #1592 .