Demo: Generative quantum advantage (#1567)
**Demo: Generative quantum advantage for classical and quantum
problems:**

**Summary:**
This is a new demo for this paper:
[https://arxiv.org/abs/2509.09033](https://arxiv.org/abs/2509.09033)

It is a quantum machine learning paper. The demo is rather critical,
since the main aim of the demo is to draw attention to the fact that
their approach is in a sense 'cheating' given the common understanding
of the term 'learning' in ML, and that the assumptions they use are not
at all common in real-world ML.

The text is hopefully fairly low on sass however. I have tried to be as
matter of fact as possible whilst still calling them out where
necessary. The most opinionated section is the last, where I try to
convince the reader that what they do is not ultimately useful for QML.

**Relevant references:**
https://arxiv.org/abs/2509.09033

**Possible Drawbacks:**

**Related GitHub Issues:**

----

* GOALS — Why are we working on this now?
This paper was released recently by Google/Caltech and has and will get
a lot of attention as an important result in the field. Publishing a
critical demo like this is part of our 'thought leader' mission.

* AUDIENCE — Who is this for?

PhD+ researchers in quantum computing. 

* KEYWORDS — What words should be included in the marketing post?

quantum machine learning
generative AI
complexity

* Which of the following types of documentation is most similar to your
file?
(more details
[here](https://www.notion.so/xanaduai/Different-kinds-of-documentation-69200645fe59442991c71f9e7d8a77f8))
    
   - [x] Demo

---------

Co-authored-by: Daniela Angulo <42325731+daniela-angulo@users.noreply.github.com>
Co-authored-by: Vasilis Belis <48251467+vbelis@users.noreply.github.com>
Co-authored-by: Josh Izaac <josh146@gmail.com>